+++ src/services/snack.ts
@@ -1,5 +1,4 @@
 import { mdiAlertCircle, mdiCheckCircleOutline } from '@mdi/js';
-import { snackbarModule } from '@/store';
 import type { TSnack } from '@/types';
 
 export const snackSuccess = async ({ message = 'Success', icon = mdiCheckCircleOutline, timeout=5000, loading=false }: TSnack): Promise<void> => {