+++ src/store/loading.ts
@@ -1,7 +1,7 @@
 import { defineStore } from 'pinia';
 import { ModuleName } from '@/types/enum_module';
 
-export const loadingStore = defineStore(ModuleName.LOADING, {
+export const loadingModule = defineStore(ModuleName.LOADING, {
 
 	state: () => ({
 		loading: false,